ref(profiling): rename profilingOnboardingSidebar
