+++ static/app/components/profiling/ProfilingOnboarding/profilingOnboardingSidebar.tsx
